server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2
bcrypt
PyJWT
python-multipart
cloudinary
pydantic_settings